requirements.txt
dateparser==1.1.1
pandas==1.4.2
plotly==5.6.0
prometheus_api_client==0.5.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
schema==0.7.5
